+++ src/docs.jl
@@ -28,4 +28,5 @@ const OnCollisionDetected = new_event_type!(registry, :OnCollisionDetected)
 world = World(Position, Velocity, Altitude, Health, ChildOf)
 
 parent = new_entity!(world, (Position(0, 0),))
+parent2 = new_entity!(world, (Position(0, 0),))
 entity = new_entity!(world, (Position(0, 0), Velocity(0, 0), ChildOf()); relations=(ChildOf => parent,))