+++ test/test_world.jl
@@ -586,6 +586,82 @@ end
     )
 end
 
+@testset "World set relations batch" begin
+    world = World(
+        Dummy,
+        Position,
+        ChildOf,
+        ChildOf2,
+    )
+    parent1 = new_entity!(world, ())
+    parent2 = new_entity!(world, ())
+    parent3 = new_entity!(world, ())
+    parent4 = new_entity!(world, ())
+    dead_parent = new_entity!(world, ())
+    remove_entity!(world, dead_parent)
+
+    filter1 = Filter(world, (ChildOf,); relations=(ChildOf => parent1,), register=true)
+    filter2 = Filter(world, (ChildOf,); register=true)
+
+    e1 = new_entity!(world, (Position(1, 1), ChildOf()); relations=(ChildOf => parent1,))
+    e2 = new_entity!(world, (Position(2, 2), ChildOf()); relations=(ChildOf => parent1,))
+    e3 = new_entity!(world, (Position(3, 3), ChildOf()); relations=(ChildOf => parent2,))
+    e4 = new_entity!(world, (Position(4, 4), ChildOf()); relations=(ChildOf => parent2,))
+    e5 = new_entity!(world, (Position(5, 5), ChildOf()); relations=(ChildOf => parent3,))
+    e6 = new_entity!(world, (Position(6, 6), ChildOf()); relations=(ChildOf => parent3,))
+
+    count = 0
+    set_relations!(world, filter1, (ChildOf => parent2,)) do entities
+        @test length(entities) == 2
+        count += 1
+    end
+    @test count == 1
+
+    @test get_relations(world, e1, (ChildOf,)) == (parent2,)
+    @test get_relations(world, e2, (ChildOf,)) == (parent2,)
+
+    count = 0
+    total = 0
+    set_relations!(world, Filter(world, (ChildOf,)), (ChildOf => parent1,)) do entities
+        count += 1
+        total += length(entities)
+    end
+    @test count == 2
+    @test total == 6
+    for (i, e) in enumerate([e1, e2, e3, e4, e5, e6])
+        @test get_relations(world, e, (ChildOf,)) == (parent1,)
+        @test get_components(world, e, (Position,)) == (Position(i, i),)
+    end
+
+    count = 0
+    set_relations!(world, Filter(world, (ChildOf,); relations=(ChildOf => parent2,)), (ChildOf => parent1,)) do entities
+        count += 1
+    end
+    @test count == 0
+
+    count = 0
+    set_relations!(world, filter1, (ChildOf => parent1,)) do entities
+        count += 1
+    end
+    @test count == 0
+
+    set_relations!(world, filter2, (ChildOf => parent4,))
+    for (i, e) in enumerate([e1, e2, e3, e4, e5, e6])
+        @test get_relations(world, e, (ChildOf,)) == (parent4,)
+        @test get_components(world, e, (Position,)) == (Position(i, i),)
+    end
+    @test world._targets[parent4._id] == true
+
+    @test_throws(
+        "ArgumentError: can't use a dead entity as relation target, except for the zero entity",
+        set_relations!(world, filter2, (ChildOf => dead_parent,))
+    )
+
+    query = Query(world, (ChildOf,))
+    @test count_entities(query) == 6
+    close!(query)
+end
+
 @testset "World copy_entity!" begin
     world = World(
         Dummy,